# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Bakery POS API documentation",
        "title": "Bakery POS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
...
        "/cart/{product_id}": {
            "put": {
                "description": "Increase or decrease quantity of a product in the user's cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Update product quantity in cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quantity change",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FormEditCart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CartItemResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CartItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImageResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "sale_price": {
                    "type": "number"
                }
            }
        },
...
        "models.FormEditCart": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "models.ImageResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "public_url": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
    }
}